adding an if statement this if test if the equipment picture is not the default do it will delete the picture from the storage

+++ lib/features/Equipments/services/db_service.dart
@@ -91,16 +91,19 @@ class DatabaseService {
   void deleteEquipment({
     required String idEquipment,
     required String tagName,
+    required String photoURL,
   }) {
     FirebaseFirestore.instance
         .collection(tagNamesCollectionRef)
         .doc(tagName)
         .delete();
-    Reference equipmentPictureRef = FirebaseStorage.instance
-        .ref()
-        .child(equipmnetPictureDic)
-        .child("${tagName}_equipment_picture");
-    equipmentPictureRef.delete();
+    if (photoURL != defaultEquipmentPicture) {
+      Reference equipmentPictureRef = FirebaseStorage.instance
+          .ref()
+          .child(equipmnetPictureDic)
+          .child("${tagName}_equipment_picture");
+      equipmentPictureRef.delete();
+    }
     _equipmentsRef.doc(idEquipment).delete();
   }
